% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateEntropy.R
\name{CalculateEntropy}
\alias{CalculateEntropy}
\title{Calculate the entropy weight for the cGEPs - Version 9 - UCASpatial}
\usage{
CalculateEntropy(object, features, assay = "RNA", slot = "data", unit = "log2")
}
\arguments{
\item{object}{reference single cell RNA-seq data: Seurat object}

\item{features}{pre-calculated marker gene dataframe by Seurat FindAllMarkers, which should include p_val, avg_logFC/avg_log2FC}

\item{assay}{optional: set the default assay of object for calculating marker weights. By default is 'RNA'.}

\item{slot}{optional: set the default slot of object for calculating marker weights. By default is 'data'.}

\item{unit}{optional: method for calculate entropy. By default is 'log2'.}
}
\value{
a list including 1: simulated spatial transcriptomics expression matrix; 2: spot-cell-type composition
}
\description{
This function can simulate a Spatial data set with n spots using single cell data.
}
