% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTopicCellType.R
\name{dot_plot_profiles_fun}
\alias{dot_plot_profiles_fun}
\title{Draw the dot plot for the nmf step}
\usage{
dot_plot_profiles_fun(h, train_cell_clust)
}
\arguments{
\item{h}{an input data after UCASpatial deconv. For example, "UCASpatial_result[[1]][[1]]@h"}

\item{train_cell_clust}{an input data after UCASpatial deconv. For example, "UCASpatial_result[[1]][[2]]"}
}
\value{
a list including two ggplot2 object: 1. dot plot for topic-each cell; 2. dot plot for topic-each cell type.
}
\description{
This function can draw the dot plot the nmf step to evalute the feature extraction process.
}
