% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewSimulate.R
\name{new_simulate}
\alias{new_simulate}
\title{Generate simulated spatial transcriptomics data - UCASpatial}
\usage{
new_simulate(
  sc_ref,
  clust_vr,
  n = 1000,
  verbose = TRUE,
  max.celltype = 4,
  max.cell = 5,
  min.celltype = 1,
  min.cell = 1
)
}
\arguments{
\item{sc_ref}{reference single cell RNA-seq data: Seurat object}

\item{clust_vr}{colname of the meta.data in sc_ref which include the cluster information}

\item{n}{optional: set the spots number for generation. By default is '1000'.}

\item{verbose}{By default is 'TRUE'.}

\item{max.celltype}{optional: set the max cell type included for a spot (randomly select from min~max). By default is '4'.}

\item{max.cell}{optional: set the max cell number included for a spot (randomly select from min~max). By default is '5'.}

\item{min.celltype}{optional: set the min cell type included for a spot (randomly select from min~max). By default is '1'.}

\item{min.cell}{optional: set the min cell number included for a spot (randomly select from min~max). By default is '1'.}
}
\value{
a list including 1: simulated spatial transcriptomics expression matrix; 2: spot-cell-type composition
}
\description{
This function can simulate a Spatial data set with n spots using single cell data.
}
