% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSpatialScatterpie.R
\name{spatial_scatterpie}
\alias{spatial_scatterpie}
\title{Draw the scatterpie plot for spatial transcriptomics data}
\usage{
spatial_scatterpie(
  st_vis,
  cell_types_all,
  img_path,
  cols = NULL,
  cell_types_interest = NULL,
  slice = NULL,
  scatterpie_alpha = 1,
  pie_scale = 1
)
}
\arguments{
\item{st_vis}{spatial transcriptomic data: Seurat object}

\item{cell_types_all}{a vecter or list data includes the cell type information}

\item{img_path}{the corresponding image path for st_vis}

\item{cols}{optional: set the color for each cell type in the pie plot. By default is 'NULL'.}

\item{cell_types_interest}{optional: Used cell types for drawing the pie plot. By default is 'NULL', which means use all cell types.}

\item{slice}{optional: set the slice used for drawing the pie plot. By default is 'NULL'.}

\item{scatterpie_alpha}{optional: set the diaphaneity/alpha of the pie. By default is '1'.}

\item{pie_scale}{optional: set the size of the pie. By default is '1'.}
}
\value{
a ggplot2 object
}
\description{
This function can draw the scatterpie plot for spatial transcriptomics data.
}
