% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_Metrics_InSilico.R
\name{test_MultiIndex}
\alias{test_MultiIndex}
\title{Test Performance In-Silico with Multi-metrics}
\usage{
test_MultiIndex(deconv_result, synthetic_comp)
}
\description{
This function can test the performance of deconvlution methods using: Pearson correlation, Spearman correlation, F1 score, RMSE, and so on.
}
