% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCASpatialDeconv.R
\name{UCASpatial_deconv}
\alias{UCASpatial_deconv}
\title{Ultra-high resolution Cellular annotation of Spatial transcriptomics - UCASpatial}
\usage{
UCASpatial_deconv(
  sc_ref,
  st_vis,
  clust_vr,
  assay = "RNA",
  slot = "data",
  output_path = NULL,
  cluster_markers = NULL,
  min.pct = 0.2,
  logfc.threshold = 0.25,
  min.diff.pct = 0.1,
  normalize = "uv",
  downsample_n = 1,
  n_cluster = 100,
  n_top = NULL,
  remove.RPL = F,
  remove.MT = F,
  cos.filter = T,
  cos.mu = 1,
  cos.n_genes_user = 900,
  marker.slot = "data",
  min_cont = 0.01,
  unit = "log2",
  random.seed = 10000,
  meta.filter = T,
  nmf.tol = 1e-04,
  meta.assay = "integrated",
  meta.ndims = 30,
  meta.resolution = 100,
  meta.purity = 0.95,
  ent.filter.threshold = 0.5,
  cos.filter.threshold = 0.05,
  weight.filter.threshold = 0.2
)
}
\arguments{
\item{sc_ref}{reference single cell RNA-seq data: Seurat object}

\item{st_vis}{spatial transcriptomic data: Seurat object}

\item{clust_vr}{colname of the meta.data in sc_ref which include the cluster information}

\item{assay}{optional: set the default assay for sc_ref. By default is 'RNA'.}

\item{slot}{optional: set the default slot for sc_ref. By default is 'counts'.}

\item{cluster_markers}{optional: give the marker genes for each cluster of sc_ref. If is NULL, it will be calculated use the default parameters. By default is 'NULL'.}

\item{normalize}{optional: method for normalizing the count matrix, including uv: unit variance, raw: no transformation applied. By default is 'uv'.}

\item{n_cluster}{optional: set the maximum size for each cluster after downsample. By default is '100'.}

\item{n_top}{optional: set the n_top marker gene used for each cluster. By default is 'NULL'.}

\item{min_cont}{optional: set the cutoff threshold. By default is '0.01'.}
}
\description{
This package could deconvolute the spatial transcriptomics data into cell-type-composition based on a reference scRNA-seq data with appropriate cell-type labels.
}
