% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCASpatialDeconv.R
\name{UCASpatial_deconv}
\alias{UCASpatial_deconv}
\title{Ultra-high resolution Cellular annotation of Spatial transcriptomics - UCASpatial}
\usage{
UCASpatial_deconv(
  sc_ref,
  st_vis,
  clust_vr,
  assay = "RNA",
  slot = "data",
  output_path = NULL,
  cluster_markers = NULL,
  min.pct = 0.2,
  logfc.threshold = 0.25,
  min.diff.pct = 0.1,
  normalize = "uv",
  downsample_n = 1,
  n_cluster = 100,
  n_top = NULL,
  remove.RPL = F,
  remove.MT = F,
  cos.filter = T,
  cos.mu = 1,
  cos.n_genes_user = 900,
  marker.slot = "data",
  min_cont = 0.01,
  unit = "log2",
  random.seed = 10000,
  meta.filter = T,
  nmf.tol = 1e-04,
  meta.assay = "integrated",
  meta.ndims = 30,
  meta.resolution = 100,
  meta.purity = 0.95,
  ent.filter.threshold = 0.5,
  cos.filter.threshold = 0.05,
  weight.filter.threshold = 0.2
)
}
\arguments{
\item{sc_ref}{reference single cell RNA-seq data: Seurat object}

\item{st_vis}{spatial transcriptomic data: Seurat object}

\item{clust_vr}{colname of the meta.data in sc_ref which include the cluster information}

\item{assay}{optional: set the default assay for sc_ref. By default is 'RNA'.}

\item{slot}{optional: set the default slot for sc_ref. By default is 'counts'.}

\item{output_path}{optional: set the output path for the intermediate files. If is NULL, it will use current path by 'getwd()' function. By default is 'NULL'.}

\item{cluster_markers}{optional: give the marker genes for each cluster of sc_ref. If is NULL, it will be calculated use the default parameters. By default is 'NULL'.}

\item{min.pct}{optional: parameter for Seurat::FindAllMarkers}

\item{logfc.threshold}{optional: parameter for Seurat::FindAllMarkers}

\item{min.diff.pct}{optional: parameter for Seurat::FindAllMarkers}

\item{normalize}{optional: method for normalizing the count matrix, including uv: unit variance, raw: no transformation applied. By default is 'uv'.}

\item{downsample_n}{optional: set the downsample number of times. By default is '1'.}

\item{n_cluster}{optional: set the maximum size for each cluster after downsample. By default is '100'.}

\item{n_top}{optional: set the n_top marker gene used for each cluster. By default is 'NULL'.}

\item{remove.RPL}{whether remove the RPL-related genes. By default is 'FALSE'.}

\item{remove.MT}{whether remove the MT-related genes. By default is 'FALSE'.}

\item{cos.filter}{optional: whether use the cosg methods to filter some features. By default is 'TRUE'.}

\item{cos.mu}{optional: parameter for cosg.}

\item{cos.n_genes_user}{optional: parameter for cosg.}

\item{marker.slot}{optional: set the slot of sc_ref for marker finding. By default is 'data'.}

\item{min_cont}{optional: parameter for calculating entropy. By default is '0.01'.}

\item{unit}{optional: parameter for calculating entropy. By default is 'log2'.}

\item{random.seed}{optional: set the random seed for randomly downsampling. By default is '10000'.}

\item{meta.filter}{optional: whether execute meta-purity filtering process. By default is 'TRUE'.}

\item{nmf.tol}{optional: parameter for the nmf iteration. By default is '1e-04'.}

\item{meta.assay}{optional: assay parameter for meta-purity filtering. By default is 'integrated'.}

\item{meta.ndims}{optional: ndims parameter for meta-purity filtering. By default is '30'.}

\item{meta.resolution}{optional: resolution parameter for meta-purity filtering. By default is '100'.}

\item{meta.purity}{optional: purity parameter for meta-purity filtering. By default is '0.95'.}

\item{ent.filter.threshold}{optional: parameter for filter low quality features based on entropy.}

\item{cos.filter.threshold}{optional: parameter for filter low quality features based on cosine similarity.}

\item{weight.filter.threshold}{optional: parameter for filter low quality features based on weight.}
}
\value{
a list including NMF result and deconvolution result
}
\description{
This package could deconvolute the spatial transcriptomics data into cell-type-composition based on a reference scRNA-seq data with appropriate cell-type labels.
}
